+++ src/features/translation/ui/translation.tsx
@@ -1,12 +1,4 @@
 import {
-  useState,
-  ChangeEvent,
-  KeyboardEvent,
-  ClipboardEvent,
-  useCallback,
-} from "react";
-
-import {
   Box,
   Button,
   Textarea,
@@ -16,81 +8,35 @@ import {
   Heading,
   Flex,
 } from "@chakra-ui/react";
+import { ChangeEvent, KeyboardEvent, ClipboardEvent } from "react";
 
-import axios from "axios";
-import { MODELS, DEFAULT_MODEL } from "@/utils";
+type TranslationPresenterProps = {
+  input: string;
+  english: string;
+  vietnamese: string;
+  selectedModel: string;
+  isLoading: boolean;
+  models: string[];
+  onInputChange: (e: ChangeEvent<HTMLTextAreaElement>) => void;
+  onKeyDown: (e: KeyboardEvent<HTMLTextAreaElement>) => void;
+  onPaste: (e: ClipboardEvent<HTMLTextAreaElement>) => void;
+  onModelChange: (e: ChangeEvent<HTMLSelectElement>) => void;
+  onTranslate: (text: string) => void;
+};
 
-export const Translation = () => {
-  const [input, setInput] = useState("");
-  const [english, setEnglish] = useState("");
-  const [vietnamese, setVietnamese] = useState("");
-  const [selectedModel, setSelectedModel] = useState(DEFAULT_MODEL);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleTranslate = useCallback(
-    async (textToTranslate: string) => {
-      if (!textToTranslate.trim()) {
-        setInput("");
-        setEnglish("");
-        setVietnamese("");
-        return;
-      }
-
-      setIsLoading(true);
-      setEnglish("Loading...");
-      setVietnamese("Loading...");
-      try {
-        const res = await axios.post("/api/translate", {
-          japaneseText: textToTranslate,
-          model: selectedModel,
-        });
-        const { english: translatedEnglish, vietnamese: translatedVietnamese } =
-          res.data.result || {};
-        if (translatedEnglish && translatedVietnamese) {
-          setEnglish(translatedEnglish);
-          setVietnamese(translatedVietnamese);
-        } else {
-          setEnglish("");
-          setVietnamese("");
-          console.error("Translation data is missing or incomplete.");
-        }
-      } catch (error) {
-        console.error("Translation Error:", error);
-        setEnglish(
-          "Error occurred during translation. Please try again later or choose another model."
-        );
-        setVietnamese(
-          "Error occurred during translation. Please try again later or choose another model."
-        );
-      } finally {
-        setIsLoading(false);
-      }
-    },
-    [selectedModel]
-  );
-
-  const handleModelChange = (e: ChangeEvent<HTMLSelectElement>) => {
-    setSelectedModel(e.target.value);
-  };
-
-  const handleKeyDown = (e: KeyboardEvent<HTMLTextAreaElement>) => {
-    if (e.key === "Enter" && (e.ctrlKey || e.metaKey)) {
-      e.preventDefault();
-      handleTranslate(input);
-    }
-  };
-
-  const handlePaste = (e: ClipboardEvent<HTMLTextAreaElement>) => {
-    e.preventDefault();
-    const pastedText = e.clipboardData.getData("text");
-    setInput(pastedText);
-    handleTranslate(pastedText);
-  };
-
-  const handleInputChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
-    setInput(e.target.value);
-  };
-
+export const Translation = ({
+  input,
+  english,
+  vietnamese,
+  selectedModel,
+  isLoading,
+  models,
+  onInputChange,
+  onKeyDown,
+  onPaste,
+  onModelChange,
+  onTranslate,
+}: TranslationPresenterProps) => {
   return (
     <Box padding="5">
       <Heading as="h1" size="lg" mb="4">
@@ -101,9 +47,9 @@ export const Translation = () => {
       </Heading>
       <Textarea
         value={input}
-        onChange={handleInputChange}
-        onKeyDown={handleKeyDown}
-        onPaste={handlePaste}
+        onChange={onInputChange}
+        onKeyDown={onKeyDown}
+        onPaste={onPaste}
         placeholder="Enter your Japanese paragraph here"
         height="xs"
         mb="4"
@@ -112,13 +58,8 @@ export const Translation = () => {
         Paste or press Ctrl+Enter (or Cmd+Enter) to translate
       </Text>
       <Flex align="center" mb="4" w="30%">
-        <Select
-          value={selectedModel}
-          onChange={handleModelChange}
-          flex="1"
-          mr="4"
-        >
-          {MODELS.map((model) => (
+        <Select value={selectedModel} onChange={onModelChange} flex="1" mr="4">
+          {models.map((model) => (
             <option key={model} value={model}>
               {model}
             </option>
@@ -126,7 +67,7 @@ export const Translation = () => {
         </Select>
         <Button
           colorScheme="blackAlpha"
-          onClick={() => handleTranslate(input)}
+          onClick={() => onTranslate(input)}
           disabled={isLoading}
         >
           Translate
